# Prologis.Workshop.ConsoleClient/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Workshop Demo",
    "description": "Nur zur Demo im Workshop",
    "contact": {
      "name": "Joerg Krause",
      "email": "joerg@joergkrause.de"
    },
    "license": {
      "name": "No License"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Label": {
      "get": {
        "tags": [
          "Label"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabelDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Label"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LabelDto"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Client Error"
          },
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Label/{id}": {
      "get": {
        "tags": [
          "Label"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabelDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Label"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LabelDto"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Client Error"
          },
          "200": {
            "description": "Success"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "delete": {
        "tags": [
          "Label"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "202": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Label/bytrackingid/{trackingId}": {
      "get": {
        "tags": [
          "Label"
        ],
        "parameters": [
          {
            "name": "trackingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/Label/searchbyname": {
      "get": {
        "tags": [
          "Label"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabelDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "204": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LabelDto": {
        "required": [
          "address",
          "name",
          "trackingId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 25,
            "minLength": 0,
            "type": "string"
          },
          "trackingId": {
            "pattern": "\\d{10}",
            "type": "string"
          },
          "address": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "deviceId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}